Move AesBase32Wrapper to Root
